feat: Convert mod to ACSE
This has been lying around on my computer for about five months! I just now discovered this, and it seems to be in a working state.
This is most likely the last I'll be adding to this project, so if anyone wishes to use this, feel free to do so.